modified:   fileinuse_functions.py 	modified:   source_functions.py

+++ fileinuse_functions.py
@@ -1,6 +1,9 @@
 import os
 def is_file_in_use(filename):
     import os
+    if not os.path.isfile(filename):
+        return False
+
     try:
         os.rename(filename, filename)
         return False
